// src/types/bank-iso.type.ts
export enum BankISO {
  NETWORKING = 'NETWORKING',
  BALANCE_ENQUIRY = 'BALANCE_ENQUIRY',
  INTERNAL_TRANSFER = 'INTERNAL_TRANSFER',
  FOREX = 'FOREX',
  ACCOUNT_STATEMENT = 'ACCOUNT_STATEMENT',
  STANDING_ORDER_INTERNAL = 'STANDING_ORDER_INTERNAL',
  STANDING_ORDER_EXTERNAL = 'STANDING_ORDER_EXTERNAL',
  CANCEL_STANDING_ORDER = 'CANCEL_STANDING_ORDER',
  RTGS_TRANSFERS = 'RTGS_TRANSFERS',
  ACCOUNT_CREATING_API_1 = 'ACCOUNT_CREATING_API_1',
  ACCOUNT_CREATING_API_2 = 'ACCOUNT_CREATING_API_2',
}